+++ botan-bindings/CHANGELOG.md
@@ -24,6 +24,9 @@
 * PATCH: update documentation in the `Botan.Bindings.View` module.
 * PATCH: update documentation in the `Botan.Bindings.PwdHash` module.
 * PATCH: update documentation in the `Botan.Bindings.SRP6` module.
+* PATCH: update documentation in the `Botan.Bindings.BlockCipher` module.
+* BREAKING: in the `BlockCipher` module, remove `_128` from the identifiers of
+  pattern synonyms that start with `BOTAN_BLOCK_CIPHER_128`.
 
 ## 0.1.0.0 -- 2025-09-17